Add JSDoc comments
