+++ package.json
@@ -25,7 +25,7 @@
     "release": "yarn test && lerna publish"
   },
   "devDependencies": {
-    "@marcbouchenoire/eslint-config": "^2.3.0",
+    "@marcbouchenoire/eslint-config": "^2.4.2",
     "@marcbouchenoire/prettier-config": "^1.1.0",
     "@types/node": "^17.0.2",
     "eslint": "^8.5.0",
