+++ packages/site/package.json
@@ -27,7 +27,7 @@
     "prettier": "prettier --write --loglevel silent --ignore-path .eslintignore",
     "lint": "yarn lint:eslint && yarn lint:stylelint && yarn lint:tsc && yarn lint:prettier",
     "lint:eslint": "eslint '**/*.{cjs,mjs,ts,tsx}' --fix",
-    "lint:prettier": "yarn prettier '**/*.{cjs,mjs,ts,tsx,json,md,yml,scss}'",
+    "lint:prettier": "yarn prettier '**/*.{cjs,mjs,ts,tsx,json,md,yml,css}'",
     "lint:stylelint": "stylelint '**/*.css' --fix",
     "lint:tsc": "tsc --project tsconfig.json"
   },
