+++ packages/site/src/components/Symbols.tsx
@@ -14,30 +14,29 @@ import {
 import { Symbols, symbols } from "symbolist"
 import colors from "tailwindcss/colors"
 import { useKey } from "../hooks/use-key"
-import { PaginationCallbacks, usePagination } from "../hooks/use-pagination"
+import { usePagination } from "../hooks/use-pagination"
 import { useSystemTheme } from "../hooks/use-system-theme"
 import { NamedSymbol } from "../types"
 import { ColorPickerPopover } from "./ColorPickerPopover"
+import { Pagination } from "./Pagination"
 
 type IconProps = ComponentProps<"div"> & NamedSymbol
 
-interface PaginationProps extends ComponentProps<"nav">, PaginationCallbacks {
-  page: number
-  pages: number[]
-  truncation?: number
-}
-
-interface PaginationPageProps
-  extends ComponentProps<"button">,
-    Pick<PaginationCallbacks, "goToPage"> {
-  page: number
-}
-
 interface Weight {
+  /**
+   * The weight's common name.
+   */
   name: string
+
+  /**
+   * The weight's numbered value.
+   */
   value: number
 }
 
+/**
+ * Get all symbols along their names.
+ */
 function getNamedSymbols() {
   const namedSymbols: NamedSymbol[] = []
 
@@ -52,13 +51,6 @@ function getNamedSymbols() {
 }
 
 const namedSymbols = getNamedSymbols()
-
-const SYMBOLS_LENGTH = namedSymbols.length
-const SYMBOLS_PAGE_LENGTH = 42
-const DEFAULT_WEIGHT = 400
-const DEFAULT_COLOR_LIGHT = colors.zinc[700]
-const DEFAULT_COLOR_DARK = colors.zinc[100]
-
 const weights: Weight[] = [
   { name: "Ultralight", value: 100 },
   { name: "Thin", value: 200 },
@@ -71,6 +63,17 @@ const weights: Weight[] = [
   { name: "Black", value: 900 }
 ]
 
+const SYMBOLS_LENGTH = namedSymbols.length
+const SYMBOLS_PAGE_LENGTH = 42
+const DEFAULT_WEIGHT = 400
+const DEFAULT_COLOR_LIGHT = colors.zinc[700]
+const DEFAULT_COLOR_DARK = colors.zinc[100]
+
+/**
+ * A select element containing all font weights as options.
+ *
+ * @param props - A set of `select` props.
+ */
 function Select(props: ComponentProps<"select">) {
   return (
     <select {...props}>
@@ -83,6 +86,14 @@ function Select(props: ComponentProps<"select">) {
   )
 }
 
+/**
+ * A single symbol component.
+ *
+ * @param props - A set of `div` props.
+ * @param props.symbol - The symbol's character.
+ * @param props.name - The symbol's name.
+ * @param [props.className] - A list of one or more classes.
+ */
 const Symbol = memo(({ symbol, name, className, ...props }: IconProps) => {
   return (
     <div
@@ -107,132 +118,11 @@ const Symbol = memo(({ symbol, name, className, ...props }: IconProps) => {
   )
 })
 
-function PaginationPage({ page, goToPage, ...props }: PaginationPageProps) {
-  const handleClick = useCallback(() => {
-    goToPage(page)
-  }, [goToPage, page])
-
-  return (
-    <button {...props} onClick={handleClick}>
-      {page + 1}
-    </button>
-  )
-}
-
-function Pagination({
-  className,
-  page: currentPage,
-  pages,
-  goToPage,
-  goToPreviousPage,
-  goToNextPage,
-  truncation = 1,
-  ...props
-}: PaginationProps) {
-  const truncatedPages = useMemo(() => {
-    const leadingTruncation = currentPage - truncation
-    const trailingTruncation = currentPage + truncation + 1
-
-    const truncatedPages = pages
-      .filter((page) => {
-        const isFirstPage = page === 0
-        const isLastPage = page === pages.length - 1
-
-        if (currentPage <= truncation + 1) {
-          return isLastPage || page <= truncation * 2 + 2
-        } else if (currentPage >= pages.length - truncation - 2) {
-          return isFirstPage || page >= pages.length - (truncation * 2 + 3)
-        } else {
-          return (
-            isFirstPage ||
-            isLastPage ||
-            (page >= leadingTruncation && page < trailingTruncation)
-          )
-        }
-      })
-      .flatMap((page, index, truncatedPages) => {
-        const followingPage = truncatedPages[index + 1]
-
-        return followingPage && Math.abs(page - followingPage) > 1
-          ? [page, null]
-          : [page]
-      })
-
-    return truncatedPages.length === 0 ? [0] : truncatedPages
-  }, [currentPage, pages, truncation])
-  const isFirstPage = useMemo(() => currentPage === 0, [currentPage])
-  const isLastPage = useMemo(
-    () => currentPage === truncatedPages.length - 1,
-    [currentPage, truncatedPages]
-  )
-
-  return (
-    <nav
-      aria-label="Pagination"
-      className={clsx(
-        "flex gap-2 justify-center items-center h-9 text-sm",
-        className
-      )}
-      {...props}
-    >
-      <button
-        aria-label="Previous"
-        className="flex justify-center items-center w-9 h-full font-medium bg-transparent rounded-md disabled:opacity-40 transition text-zinc-500 dark:text-zinc-400 hover:bg-zinc-500/10 dark:hover:bg-zinc-100/10 disabled:!bg-transparent focusable"
-        disabled={isFirstPage}
-        onClick={goToPreviousPage}
-      >
-        <svg height="24" role="presentation" width="24">
-          <path
-            d="M9.005 10.995l4.593-4.593a.99.99 0 111.4 1.4l-3.9 3.9 3.9 3.9a.99.99 0 01-1.4 1.4L9.005 12.41a1 1 0 010-1.414z"
-            fill="currentColor"
-            fillRule="evenodd"
-          />
-        </svg>
-      </button>
-      {truncatedPages.map((page, index) =>
-        page === null ? (
-          <span
-            className="flex justify-center items-center w-9 h-full font-medium text-zinc-400 dark:text-zinc-500"
-            key={`ellipsis-${index}`}
-          >
-            …
-          </span>
-        ) : (
-          <PaginationPage
-            aria-current={currentPage === page ? "page" : false}
-            className={clsx(
-              "w-9 h-full font-medium rounded-md transition focusable",
-              {
-                "hover:shadow-primary-500/5 dark:hover:shadow-primary-400/5 selection:bg-white/30 dark:selection:bg-zinc-900/30 dark:text-zinc-900 bg-primary-500 dark:bg-primary-400 shadow-lg focusable shadow-primary-500/10 dark:shadow-primary-400/10 text-white hover:bg-opacity-80 dark:hover:bg-opacity-80":
-                  currentPage === page,
-                "hover:bg-zinc-500/10 dark:hover:bg-zinc-100/10":
-                  currentPage !== page
-              }
-            )}
-            goToPage={goToPage}
-            key={`${page}-${index}`}
-            page={page}
-          />
-        )
-      )}
-      <button
-        aria-label="Previous"
-        className="flex justify-center items-center w-9 h-full font-medium bg-transparent rounded-md disabled:opacity-40 transition text-zinc-500 dark:text-zinc-400 hover:bg-zinc-500/10 dark:hover:bg-zinc-100/10 disabled:!bg-transparent focusable"
-        disabled={isLastPage}
-        onClick={goToNextPage}
-      >
-        <svg height="24" role="presentation" width="24">
-          <path
-            d="M14.995 10.995a1 1 0 010 1.414l-4.593 4.593a.99.99 0 01-1.4-1.4l3.9-3.9-3.9-3.9a.99.99 0 011.4-1.4l4.593 4.593z"
-            fill="currentColor"
-            fillRule="evenodd"
-          />
-        </svg>
-      </button>
-    </nav>
-  )
-}
-
+/**
+ * An interactive section to explore all symbols.
+ *
+ * @param props - A set of `section` props.
+ */
 export function Symbols(props: ComponentProps<"section">) {
   const searchRef = useRef<HTMLInputElement>(null)
   const [search, setSearch] = useState("")
