+++ packages/site/src/hooks/use-key.ts
@@ -1,9 +1,12 @@
 import { useCallback, useEffect } from "react"
 
-export const useKey = (
-  key: string,
-  callback: (event?: KeyboardEvent) => void
-) => {
+/**
+ * Invoke a function when pressing a specific key.
+ *
+ * @param key - The key to press.
+ * @param callback - The function to invoke when pressing the key.
+ */
+export function useKey(key: string, callback: (event?: KeyboardEvent) => void) {
   const handleKeyDown = useCallback(
     (event: KeyboardEvent) => {
       if (event.key === key) {
