+++ packages/site/src/pages/_document.tsx
@@ -1,6 +1,12 @@
 import NextDocument, { Head, Html, Main, NextScript } from "next/document"
 
+/**
+ * A custom `Document` component.
+ */
 class Document extends NextDocument {
+  /**
+   * Describe the document markup.
+   */
   render() {
     return (
       <Html lang="en">
