+++ packages/site/src/plugins/rehype/remove-images.ts
@@ -1,6 +1,9 @@
 import { Plugin } from "unified"
 import { remove } from "unist-util-remove"
 
+/**
+ * A plugin to remove all image nodes.
+ */
 const removeImages: Plugin = () => {
   return (tree) => {
     remove(tree, { tagName: "img", type: "element" })
