+++ packages/site/src/types.ts
@@ -1,4 +1,11 @@
 export interface NamedSymbol {
+  /**
+   * The symbol's name.
+   */
   name: string
+
+  /**
+   * The symbol's character.
+   */
   symbol: string
 }
