+++ packages/symbolist/.eslintrc.json
@@ -1,3 +1,6 @@
 {
-  "extends": ["@marcbouchenoire/eslint-config"]
+  "extends": [
+    "@marcbouchenoire/eslint-config",
+    "@marcbouchenoire/eslint-config/jsdoc"
+  ]
 }
