+++ packages/symbolist/scripts/utils/is-macOS.ts
@@ -1,3 +1,6 @@
+/**
+ * Whether the current platform is macOS.
+ */
 export function isMacOS() {
   return process.platform === "darwin"
 }
