+++ packages/symbolist/scripts/utils/listr-clipboard.ts
@@ -2,6 +2,12 @@ import clipboard from "clipboardy"
 import { ListrTask } from "listr"
 import { ListrInput } from "./listr-input"
 
+/**
+ * A Listr input using the clipboard's content.
+ *
+ * @param title - The action's message.
+ * @param callback - The function to invoke when submitting.
+ */
 export function ListrClipboard<C>(
   title: string,
   callback: (value: string, context: C) => void
