+++ packages/symbolist/scripts/utils/listr-input.ts
@@ -3,6 +3,16 @@ import ListrDefaultInput from "listr-input"
 import pc from "picocolors"
 import { join } from "./join"
 
+/**
+ * A Listr input.
+ *
+ * @param title - The action's message.
+ * @param placeholder - The input's placeholder.
+ * @param callback - The function to invoke when submitting.
+ * @param [message] - The action's secondary message.
+ * @param [warning] - The action's warning.
+ * @param [autoSubmit] - A function to decide whether to submit the current value.
+ */
 export function ListrInput<C>(
   title: string,
   placeholder: string,
