+++ packages/symbolist/scripts/utils/silent-error.ts
@@ -1,6 +1,14 @@
 const SILENT_ERROR_NAME = "SilentError"
 
+/**
+ * An `Error` that can be filtered and silenced.
+ */
 export class SilentError extends Error {
+  /**
+   * Create a `SilentError` with a message.
+   *
+   * @param message - The error message.
+   */
   constructor(message: string) {
     super(message)
 
@@ -8,6 +16,11 @@ export class SilentError extends Error {
   }
 }
 
+/**
+ * Whether the error is a `SilentError`.
+ *
+ * @param error - The error to check.
+ */
 export function isSilentError(
   error: Error | SilentError
 ): error is SilentError {
