+++ packages/symbolist/tests/helpers.ts
@@ -2,9 +2,15 @@ import { Test, suite } from "uvu"
 
 type Describer = (test: Test) => Promise<void> | void
 
-export function describe(name: string, hook: Describer): void {
+/**
+ * Run multiple tests as a named suite.
+ *
+ * @param name - The name of the suite.
+ * @param callback - The suite as a function.
+ */
+export function describe(name: string, callback: Describer): void {
   const test = suite(name)
-  hook(test)
+  callback(test)
 
   test.run()
 }
